refactor: reduce boilerplate for remaining docs
