+++ src/pages/boundary/docs/[[...page]].tsx
@@ -2,7 +2,7 @@ import DocsView from 'views/docs-view'
 import { getRootDocsPathGenerationFunctions } from 'views/docs-view/utils/get-root-docs-path-generation-functions'
 
 const { getStaticPaths, getStaticProps } = getRootDocsPathGenerationFunctions(
-	'consul',
+	'boundary',
 	'docs'
 )
